+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+MINOR
+- Add default implementation of CanvasElement with basically initialized fields
+- Add default mouse and scroll event streams
+
 # Changelog
 
 ## 1.5.6